# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LocalMarket API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: ads
    description: Ads/Publicités
  - name: admin
    description: Admin operations
  - name: stats
    description: Statistics
  - name: auth
    description: User authentication
  - name: categories
    description: Categories
  - name: units
    description: Units of measure
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /ads:
    get:
      operationId: listAds
      tags: [ads]
      summary: List published ads
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: location
          in: query
          schema:
            type: string
        - name: product
          in: query
          schema:
            type: string
        - name: quantity
          in: query
          schema:
            type: string
        - name: unit
          in: query
          schema:
            type: string
        - name: listingType
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"
    post:
      operationId: createAd
      tags: [ads]
      summary: Submit a new ad
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdInput"
      responses:
        "201":
          description: Created ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"

  /ads/{id}:
    get:
      operationId: getAd
      tags: [ads]
      summary: Get ad by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get platform statistics
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

  /categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List active categories
      responses:
        "200":
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

  /units:
    get:
      operationId: listUnits
      tags: [units]
      summary: List active units
      responses:
        "200":
          description: List of units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"

  /admin/ads:
    get:
      operationId: adminListAds
      tags: [admin]
      summary: Admin - list all ads (including pending)
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: All ads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ad"

  /admin/ads/{id}/status:
    patch:
      operationId: updateAdStatus
      tags: [admin]
      summary: Admin - update ad status
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdStatusUpdate"
      responses:
        "200":
          description: Updated ad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Ad"

  /admin/ads/{id}:
    delete:
      operationId: deleteAd
      tags: [admin]
      summary: Admin - delete ad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /admin/branding:
    get:
      operationId: getBranding
      tags: [admin]
      summary: Get current branding settings
      responses:
        "200":
          description: Branding settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branding"
    put:
      operationId: updateBranding
      tags: [admin]
      summary: Update branding settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Branding"
      responses:
        "200":
          description: Updated branding
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Branding"

  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin]
      summary: Admin - list all categories
      responses:
        "200":
          description: All categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
    post:
      operationId: adminCreateCategory
      tags: [admin]
      summary: Admin - create category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

  /admin/categories/{id}:
    put:
      operationId: adminUpdateCategory
      tags: [admin]
      summary: Admin - update category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: Updated category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
    delete:
      operationId: adminDeleteCategory
      tags: [admin]
      summary: Admin - delete category
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /admin/units:
    get:
      operationId: adminListUnits
      tags: [admin]
      summary: Admin - list all units
      responses:
        "200":
          description: All units
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unit"
    post:
      operationId: adminCreateUnit
      tags: [admin]
      summary: Admin - create unit
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "201":
          description: Created unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"

  /admin/units/{id}:
    put:
      operationId: adminUpdateUnit
      tags: [admin]
      summary: Admin - update unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnitInput"
      responses:
        "200":
          description: Updated unit
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unit"
    delete:
      operationId: adminDeleteUnit
      tags: [admin]
      summary: Admin - delete unit
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /admin/promotion-prices:
    get:
      operationId: adminListPromotionPrices
      tags: [admin]
      summary: Admin - list promotion prices
      responses:
        "200":
          description: All promotion prices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromotionPrice"
    post:
      operationId: adminCreatePromotionPrice
      tags: [admin]
      summary: Admin - create promotion price
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionPriceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromotionPrice"

  /admin/promotion-prices/{id}:
    put:
      operationId: adminUpdatePromotionPrice
      tags: [admin]
      summary: Admin - update promotion price
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PromotionPriceInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromotionPrice"
    delete:
      operationId: adminDeletePromotionPrice
      tags: [admin]
      summary: Admin - delete promotion price
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "409":
          description: Email already in use

  /auth/login:
    post:
      operationId: userLogin
      tags: [auth]
      summary: Login a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current authenticated user
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCredentials"
      responses:
        "200":
          description: Login success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Unauthorized

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Ad:
      type: object
      required: [id, title, location, product, category, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: string
        product:
          type: string
        quantity:
          type: ["string", "null"]
        unit:
          type: ["string", "null"]
        category:
          type: string
        listingType:
          type: string
          enum: [free, flexible, fixed]
        price:
          type: ["string", "null"]
        isPromoted:
          type: boolean
        promotionDuration:
          type: ["integer", "null"]
        promotionPrice:
          type: ["string", "null"]
        contactPhone:
          type: ["string", "null"]
        contactEmail:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, published, rejected]
        createdAt:
          type: string

    AdInput:
      type: object
      required: [title, location, product, category]
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        product:
          type: string
        quantity:
          type: string
        unit:
          type: string
        category:
          type: string
        listingType:
          type: string
          enum: [free, flexible, fixed]
        price:
          type: string
        isPromoted:
          type: boolean
        promotionDuration:
          type: integer
        promotionPrice:
          type: string
        contactPhone:
          type: string
        contactEmail:
          type: string

    AdStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [pending, published, rejected]

    Stats:
      type: object
      required: [totalAds, totalExchangers, satisfaction]
      properties:
        totalAds:
          type: integer
        totalExchangers:
          type: integer
        satisfaction:
          type: number

    Branding:
      type: object
      properties:
        logoUrl:
          type: ["string", "null"]
        primaryColor:
          type: string
        accentColor:
          type: string
        backgroundColor:
          type: string
        fontFamily:
          type: string
        siteName:
          type: string

    Category:
      type: object
      required: [id, name, slug, active]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        active:
          type: boolean

    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        active:
          type: boolean

    Unit:
      type: object
      required: [id, name, symbol, active]
      properties:
        id:
          type: integer
        name:
          type: string
        symbol:
          type: string
        active:
          type: boolean

    UnitInput:
      type: object
      required: [name, symbol]
      properties:
        name:
          type: string
        symbol:
          type: string
        active:
          type: boolean

    PromotionPrice:
      type: object
      required: [id, duration, label, price, active]
      properties:
        id:
          type: integer
        duration:
          type: integer
        label:
          type: string
        price:
          type: string
        active:
          type: boolean

    PromotionPriceInput:
      type: object
      required: [duration, label, price]
      properties:
        duration:
          type: integer
        label:
          type: string
        price:
          type: string
        active:
          type: boolean

    AdminCredentials:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string

    LoginResult:
      type: object
      required: [success, token]
      properties:
        success:
          type: boolean
        token:
          type: string

    RegisterInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    UserProfile:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        createdAt:
          type: string

    AuthResult:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"

  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT