# lib/api-zod/src/generated/types/adInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalMarket API
 * OpenAPI spec version: 0.1.0
 */
import type { AdInputListingType } from './adInputListingType';

export interface AdInput {
  title: string;
  description?: string;
  location: string;
  product: string;
  quantity?: string;
  unit?: string;
  category: string;
  listingType?: AdInputListingType;
  price?: string;
  isPromoted?: boolean;
  promotionDuration?: number;
  promotionPrice?: string;
  contactPhone?: string;
  contactEmail?: string;
}