# lib/api-zod/src/generated/types/promotionPrice.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalMarket API
 * OpenAPI spec version: 0.1.0
 */

export interface PromotionPrice {
  id: number;
  duration: number;
  label: string;
  price: string;
  active: boolean;
}