# lib/api-zod/src/generated/types/unit.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LocalMarket API
 * OpenAPI spec version: 0.1.0
 */

export interface Unit {
  id: number;
  name: string;
  symbol: string;
  active: boolean;
}